Move Add Mul Pow into types.jl
